documentation(EJ2-71942): issue resolved
