+++ code-snippet/diagram/connectors-cs52/index.js
@@ -1,80 +1,55 @@
-var connectors = [{
-        id: "connector1",
-        targetDecorator: {
-            style: {
-                strokeColor: '#6BA5D7',
-                fill: '#6BA5D7',
-                strokeWidth: 2
-            }
-        },
-        style: {
-            // Stroke color
-            strokeColor: '#6BA5D7',
-            fill: '#6BA5D7',
-            // Stroke width of the line
-            strokeWidth: 2,
-            // Line style
-            strokeDashArray: '2,2'
-        },
-        sourcePoint: {
-            x: 100,
-            y: 100
-        },
-        targetPoint: {
-            x: 200,
-            y: 200
-        },
-        segments: [{
-            type: 'Orthogonal',
-            direction: 'Right',
-            length: 50
-        }],
+var nodes = [{
+        id: 'node1',
+        width: 100,
+        height: 100,
+        offsetX: 100,
+        offsetY: 100,
     },
     {
-        id: "connector2",
-        // Set the visibility of the connector to false
-        visible: false,
-        targetDecorator: {
-            style: {
-                strokeColor: '#6BA5D7',
-                fill: '#6BA5D7',
-                strokeWidth: 2
-            }
-        },
+        id: 'node2',
+        width: 100,
+        height: 100,
+        offsetX: 100,
+        offsetY: 350,
+    },
+]
+
+
+var connectors = [{
+    id: "connector1",
+    type: 'Orthogonal',
+       style: {
+        strokeColor: '#6BA5D7',
+        fill: '#6BA5D7',
+        strokeWidth: 2
+    },
+    targetDecorator: {
         style: {
-            // Stroke color
-            strokeColor: '#6BA5D7',
             fill: '#6BA5D7',
-            // Stroke width of the line
-            strokeWidth: 2,
-            // Line style
-            strokeDashArray: '2,2'
-        },
-        sourcePoint: {
-            x: 300,
-            y: 300
-        },
-        targetPoint: {
-            x: 400,
-            y: 400
-        },
-        segments: [{
-            type: 'Orthogonal',
-            direction: 'Right',
-            length: 50
-        }],
-    }
-];
+            strokeColor: '#6BA5D7'
+        }
+    },
+    cornerRadius: 10,
+    sourceID: 'node1',
+    targetID: 'node2',
+    segments: [{
+        type: 'Orthogonal',
+        direction: 'Right',
+        length: 50
+    }],
+}]
 
 var diagram = new ej.diagrams.Diagram({
-    width: '100%',
-    height: '600px',
-    // Define the default values for all the connector.Similary we can add all the properties
-    getConnectorDefaults: (obj) => {
-        obj.type = 'Orthogonal'
-        return obj;
+    width: 1500,
+    height: 1500,
+    getConnectorDefaults: function(node) {
+        node.height = 100;
+        node.width = 100;
+        node.style.fill = '#6BA5D7';
+        node.style.strokeColor = 'white';
+        return node;
     },
     connectors: connectors,
+    nodes: nodes
 }, '#element');
-
