+++ code-snippet/diagram/connectors-cs55/index.ts
@@ -2,18 +2,16 @@
 
 
 import {
-  Diagram, SnapConstraints,ConnectorConstraints, LineRouting, DiagramConstraints, NodeModel, ConnectorModel
+  Diagram, SnapConstraints, LineRouting, DiagramConstraints, NodeModel, ConnectorModel
 } from '@syncfusion/ej2-diagrams';
 Diagram.Inject(LineRouting);
 let nodes: NodeModel[] = [
   { id: 'shape1', offsetX: 100, offsetY: 100, width: 120, height: 50 },
-  { id: 'shape2', offsetX: 350, offsetY: 300, width: 120, height: 50 },
-  { id: 'shape3', offsetX: 150, offsetY: 200, width: 120, height: 50 },
-  { id: 'shape4', offsetX: 300, offsetY: 200, width: 120, height: 50 }
+  { id: 'shape2', offsetX: 300, offsetY: 300, width: 120, height: 50 },
+  { id: 'shape3', offsetX: 150, offsetY: 200, width: 120, height: 50 }
 ];
 let connectors: ConnectorModel[] = [
-  { id: 'connector', sourceID: 'shape1', targetID: 'shape2', type: 'Orthogonal',annotations:[{offset:.7,content:' Routing \n enabled',style:{fill:"white"}}]},
-  { id: 'connector2', sourceID: 'shape1', targetID: 'shape2',annotations:[{offset:.6,content:' Routing \n disabled',style:{fill:"white"}}], type: 'Orthogonal', constraints: ConnectorConstraints.Default &~ ConnectorConstraints.InheritLineRouting }
+  { id: 'connector', sourceID: 'shape1', targetID: 'shape2', type: 'Orthogonal' }
 ];
 //initialize the diagram control
 let diagram: Diagram = new Diagram({
