+++ code-snippet/diagram/connectors-cs58/index.ts
@@ -1,37 +1,47 @@
 
 
 
-import {
-    Diagram,NodeModel,ConnectorModel,ConnectorConstraints
-} from '@syncfusion/ej2-diagrams';
-
-let node1: NodeModel = {
-    id: 'node1',
-    width: 100, height: 100,
-    offsetX: 150, offsetY: 150,
-    annotations: [{ content: 'node1' }]
+import {Diagram,ConnectorModel} from '@syncfusion/ej2-diagrams';
+let connectors: ConnectorModel = {
+    id: 'connector1',
+    // Defines the z-index value for the connector
+    zIndex: 2,
+    type: 'Straight',
+    sourcePoint: {
+        x: 300,
+        y: 100
+    },
+    targetPoint: {
+        x: 300,
+        y: 200
+    }
 };
-let node2: NodeModel = {
-    id: 'node2',
-    width: 100, height: 100,
-    offsetX: 650, offsetY: 150,
-    annotations: [{ content: 'node2' }]
+let connectors2: ConnectorModel = {
+    id: 'connector2',
+    type: 'Straight',
+    // Defines the z-index value for the connector
+    zIndex: 1,
+    sourcePoint: {
+        x: 100,
+        y: 100
+    },
+    targetPoint: {
+        x: 200,
+        y: 200
+    }
 };
-let node3: NodeModel = {
-    id: 'node3',
-    width: 100, height: 100,
-    offsetX: 490, offsetY: 290,
-    annotations: [{ content: 'node3' }]
-};
-let connector1: ConnectorModel = {
-    id: 'connector1',sourceID:"node1",targetID:"node2",
-    constraints: ConnectorConstraints.Default | ConnectorConstraints.AllowDrop
-};
-let diagram:Diagram = new Diagram({
-    width: 1500, height: 1000,
-    enableConnectorSplit:true,
-    nodes: [node1, node2, node3],
-    connectors: [connector1]
+let diagram: Diagram = new Diagram({
+    width: '100%',
+    getConnectorDefaults: (obj: ConnectorModel): ConnectorModel => {
+        obj.style.strokeColor = '#6BA5D7';
+        obj.style.fill = '#6BA5D7';
+        obj.style.strokeWidth = 2;
+        obj.targetDecorator.style.fill = '#6BA5D7';
+        obj.targetDecorator.style.strokeColor = '#6BA5D7';
+        return obj;
+    },
+    height: '600px',
+    connectors: [connectors, connectors2]
 });
 diagram.appendTo('#element');
